incorporate recent commits into pipeline testing module
